InvalidLtiProviderException moved AccessTokenRequestValidator moved MissingScopeException moved AccessTokenRequestValidator attributes role and ltiProvider
